refactor: update AgentListWithAgent import paths
Replaced imports of AgentListWithAgent from the services file to the new types file across multiple components and services, enhancing code organization and clarity.
