+++ web-app/src/components/agents/agents.tsx
@@ -1,7 +1,7 @@
 import { useTranslations } from "next-intl";
 
-import { AgentListWithAgent } from "@/lib/db/services/agentList.service";
 import { AgentWithRelations } from "@/lib/db/types/agent.types";
+import { AgentListWithAgent } from "@/lib/db/types/agentList.types";
 import { cn } from "@/lib/utils";
 
 import { AgentCard, AgentCardSkeleton } from "./agent-card";
