+++ web-app/src/lib/db/types/agentList.types.ts
@@ -1,0 +1,9 @@
+import { Prisma } from "@prisma/client";
+
+export const agentListInclude = {
+  agents: true,
+} as const;
+
+export type AgentListWithAgent = Prisma.AgentListGetPayload<{
+  include: typeof agentListInclude;
+}>;
